refactor: eleteSuggest 과외 신청 취소 기능 예외처리 추가 (#103)

+++ server/src/main/java/com/codueon/boostUp/domain/suggest/service/SuggestDbService.java
@@ -42,6 +42,11 @@ public class SuggestDbService {
             throw new BusinessLogicException(ExceptionCode.INVALID_ACCESS);
         }
 
+        if (!findSuggest.getStatus().equals(Suggest.SuggestStatus.ACCEPT_IN_PROGRESS) &&
+            !findSuggest.getStatus().equals(Suggest.SuggestStatus.PAY_IN_PROGRESS)) {
+            throw new BusinessLogicException(ExceptionCode.INVALID_ACCESS);
+        }
+
         suggestRepository.delete(findSuggest);
     }
 }
